move metricsService into services directory
