+++ src/pfe/portal/routes/projects/bind.route.js
@@ -19,7 +19,7 @@ const inflateAsync = promisify(zlib.inflate);
 const cwUtils = require('../../modules/utils/sharedFunctions');
 const Logger = require('../../modules/utils/Logger');
 const Project = require('../../modules/Project');
-const metricsService = require('../../modules/metricsService');
+const metricsService = require('../../services/metricsService');
 const ProjectInitializerError = require('../../modules/utils/errors/ProjectInitializerError');
 const { ILLEGAL_PROJECT_NAME_CHARS } = require('../../config/requestConfig');
 const router = express.Router();
